refactor(data): move all related data class to data folder
