+++ app/src/main/java/org/flyve/mdm/agent/adapter/ApplicationsAdapter.java
@@ -33,7 +33,7 @@ import android.widget.ImageView;
 import android.widget.TextView;
 
 import org.flyve.mdm.agent.R;
-import org.flyve.mdm.agent.room.entity.Application;
+import org.flyve.mdm.agent.data.database.entity.Application;
 import org.flyve.mdm.agent.utils.FlyveLog;
 import org.flyve.mdm.agent.utils.Helpers;
