+++ app/src/main/java/org/flyve/mdm/agent/adapter/FilesAdapter.java
@@ -32,7 +32,7 @@ import android.widget.BaseAdapter;
 import android.widget.TextView;
 
 import org.flyve.mdm.agent.R;
-import org.flyve.mdm.agent.room.entity.File;
+import org.flyve.mdm.agent.data.database.entity.File;
 import org.flyve.mdm.agent.utils.FlyveLog;
 
 public class FilesAdapter extends BaseAdapter {
