+++ app/src/main/java/org/flyve/mdm/agent/core/enrollment/EnrollmentModel.java
@@ -33,8 +33,8 @@ import android.support.v7.app.AlertDialog;
 import org.flyve.inventory.categories.Hardware;
 import org.flyve.mdm.agent.BuildConfig;
 import org.flyve.mdm.agent.R;
-import org.flyve.mdm.agent.data.MqttData;
-import org.flyve.mdm.agent.data.UserData;
+import org.flyve.mdm.agent.data.database.MqttData;
+import org.flyve.mdm.agent.data.localstorage.UserData;
 import org.flyve.mdm.agent.security.AndroidCryptoProvider;
 import org.flyve.mdm.agent.utils.FlyveLog;
 import org.flyve.mdm.agent.utils.Helpers;
