+++ app/src/main/java/org/flyve/mdm/agent/core/user/UserModel.java
@@ -29,7 +29,7 @@ import android.content.DialogInterface;
 import android.support.v7.app.AlertDialog;
 
 import org.flyve.mdm.agent.R;
-import org.flyve.mdm.agent.data.UserData;
+import org.flyve.mdm.agent.data.localstorage.UserData;
 import org.flyve.mdm.agent.utils.FlyveLog;
 import org.flyve.mdm.agent.utils.Helpers;
