+++ app/src/main/java/org/flyve/mdm/agent/receivers/MQTTConnectivityReceiver.java
@@ -29,7 +29,7 @@ import android.content.Intent;
 import android.telephony.PhoneStateListener;
 import android.telephony.TelephonyManager;
 
-import org.flyve.mdm.agent.data.PoliciesData;
+import org.flyve.mdm.agent.data.database.PoliciesData;
 import org.flyve.mdm.agent.policies.AirplaneModePolicy;
 import org.flyve.mdm.agent.policies.BluetoothPolicy;
 import org.flyve.mdm.agent.policies.GPSPolicy;
