+++ app/src/main/java/org/flyve/mdm/agent/services/MQTTService.java
@@ -40,8 +40,8 @@ import org.eclipse.paho.client.mqttv3.MqttConnectOptions;
 import org.eclipse.paho.client.mqttv3.MqttException;
 import org.eclipse.paho.client.mqttv3.MqttMessage;
 import org.flyve.mdm.agent.R;
-import org.flyve.mdm.agent.data.AppData;
-import org.flyve.mdm.agent.data.MqttData;
+import org.flyve.mdm.agent.data.localstorage.AppData;
+import org.flyve.mdm.agent.data.database.MqttData;
 import org.flyve.mdm.agent.policies.AirplaneModePolicy;
 import org.flyve.mdm.agent.policies.BasePolicies;
 import org.flyve.mdm.agent.policies.BluetoothPolicy;
@@ -69,7 +69,6 @@ import org.flyve.mdm.agent.policies.StatusBarPolicy;
 import org.flyve.mdm.agent.policies.StorageEncryptionPolicy;
 import org.flyve.mdm.agent.policies.StreamAccessibilityPolicy;
 import org.flyve.mdm.agent.policies.StreamAlarmPolicy;
-import org.flyve.mdm.agent.policies.StreamDTMFPolicy;
 import org.flyve.mdm.agent.policies.StreamMusicPolicy;
 import org.flyve.mdm.agent.policies.StreamNotificationPolicy;
 import org.flyve.mdm.agent.policies.StreamRingPolicy;
