+++ app/src/main/java/org/flyve/mdm/agent/ui/FragmentAppList.java
@@ -37,8 +37,8 @@ import android.widget.TextView;
 
 import org.flyve.mdm.agent.R;
 import org.flyve.mdm.agent.adapter.ApplicationsAdapter;
-import org.flyve.mdm.agent.room.database.AppDataBase;
-import org.flyve.mdm.agent.room.entity.Application;
+import org.flyve.mdm.agent.data.database.setup.AppDataBase;
+import org.flyve.mdm.agent.data.database.entity.Application;
 
 public class FragmentAppList extends Fragment {
