+++ app/src/main/java/org/flyve/mdm/agent/ui/FragmentInformation.java
@@ -39,9 +39,9 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import org.flyve.mdm.agent.R;
-import org.flyve.mdm.agent.data.SupervisorData;
-import org.flyve.mdm.agent.data.UserData;
-import org.flyve.mdm.agent.data.AppData;
+import org.flyve.mdm.agent.data.localstorage.SupervisorData;
+import org.flyve.mdm.agent.data.localstorage.UserData;
+import org.flyve.mdm.agent.data.localstorage.AppData;
 import org.flyve.mdm.agent.utils.FlyveLog;
 import org.flyve.mdm.agent.utils.Helpers;
 import org.json.JSONObject;
