+++ app/src/main/java/org/flyve/mdm/agent/ui/LockActivity.java
@@ -38,7 +38,7 @@ import android.widget.Button;
 import android.widget.TextView;
 
 import org.flyve.mdm.agent.R;
-import org.flyve.mdm.agent.data.SupervisorData;
+import org.flyve.mdm.agent.data.localstorage.SupervisorData;
 import org.flyve.mdm.agent.utils.FlyveLog;
 
 public class LockActivity extends AppCompatActivity {
