+++ app/src/main/java/org/flyve/mdm/agent/ui/MainActivity.java
@@ -39,7 +39,7 @@ import android.widget.TextView;
 
 import org.flyve.mdm.agent.R;
 import org.flyve.mdm.agent.adapter.DrawerAdapter;
-import org.flyve.mdm.agent.data.AppData;
+import org.flyve.mdm.agent.data.localstorage.AppData;
 import org.flyve.mdm.agent.services.DeviceLockedController;
 import org.flyve.mdm.agent.services.MQTTService;
 import org.flyve.mdm.agent.services.PoliciesDeviceManager;
