+++ app/src/main/java/org/flyve/mdm/agent/utils/Helpers.java
@@ -56,9 +56,9 @@ import android.view.inputmethod.InputMethodManager;
 import android.widget.TextView;
 
 import org.flyve.mdm.agent.R;
-import org.flyve.mdm.agent.data.AppData;
-import org.flyve.mdm.agent.room.database.AppDataBase;
-import org.flyve.mdm.agent.room.entity.Application;
+import org.flyve.mdm.agent.data.localstorage.AppData;
+import org.flyve.mdm.agent.data.database.setup.AppDataBase;
+import org.flyve.mdm.agent.data.database.entity.Application;
 import org.flyve.mdm.agent.ui.ErrorActivity;
 import org.flyve.mdm.agent.ui.InstallAppActivity;
 import org.flyve.mdm.agent.ui.MainActivity;
